+++ src/main/java/cn/dancingsnow/mcdrc/command/NodeType.java
@@ -1,4 +1,4 @@
-package cn.dancingsnow.mcdrssc.command;
+package cn.dancingsnow.mcdrc.command;
 
 public enum NodeType {
     LITERAL,
