+++ src/main/java/cn/dancingsnow/mcdrc/mixin/package-info.java
@@ -1,0 +1,1 @@
+package cn.dancingsnow.mcdrc.mixin;
