--- src/main/java/cn/dancingsnow/mcdrssc/mixin/package-info.java
@@ -1,1 +1,0 @@
-package cn.dancingsnow.mcdrssc.mixin;
