+++ src/main/resources/mcdrssc.mixins.json
@@ -1,7 +1,7 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "cn.dancingsnow.mcdrssc.mixin",
+  "package": "cn.dancingsnow.mcdrc.mixin",
   "compatibilityLevel": "JAVA_17",
   "mixins": [
     "MixinServerPlayNetworkHandler"
